Moving rung to flat deduplicate mmodel for consistency and space reduction. Refactor `rung`-related logic across migrations, maps, and transformers: replace `rung_id` with `rung_hash`, remove unused columns, and update merge scripts for improved consistency.

+++ src/LogixDb.Data.Sqlite/SqliteManager.cs
@@ -182,8 +182,8 @@ public sealed class SqliteManager : IDbManager
                 target,
                 transaction);
 
-            target.VersionId = result.VersionId;
-            target.VersionNumber = result.VersionNumber;
+            target.VersionId = (int)result.VersionId;
+            target.VersionNumber = (int)result.VersionNumber;
 
             // Inserts all the configured metadata for the version.
             await connection.ExecuteAsync(SqliteScript.PostInfo,
